add auth logic

+++ routers/PostsRouter.py
@@ -1,5 +1,6 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, Depends
 
+from Auth.Security import check_post_access, get_current_user
 from Schemas.PostDTO import PostAddDTO
 from Schemas.PostDTO import PostDTO
 from repos import PostRepo
@@ -10,27 +11,24 @@ router = APIRouter(
 )
 
 
-@router.get("")
+@router.get("", response_model=list[PostDTO])
 def get_posts() -> list[PostDTO]:
     return PostRepo.get_all_posts()
 
 
-@router.get("/{post_id}")
+@router.get("/{post_id}", response_model=PostDTO)
 def get_post(post_id: int) -> PostDTO:
     return PostRepo.get_post_by_id(post_id)
 
 
-@router.post("")
-def create_post(post: PostAddDTO) -> PostDTO:
-    return PostRepo.create_post(post)
-
-
-@router.put("/{post_id}")
-def update_post(post_id: int, post: PostAddDTO) -> PostDTO:
-    return PostRepo.update_post(post_id, post)
+@router.put("/{post_id}", response_model=PostDTO)
+def update_post(post_id: int, post: PostAddDTO, token_payload: dict = Depends(get_current_user)) -> PostDTO:
+    if check_post_access(post_id, token_payload=token_payload):
+        return PostRepo.update_post(post_id, post)
 
 
 @router.delete("/{post_id}")
-def delete_post(post_id: int) -> dict[str, str]:
-    PostRepo.delete_post(post_id)
-    return {"message": "post has been deleted"}
+def delete_post(post_id: int, token_payload: dict = Depends(get_current_user)) -> dict[str, str]:
+    if check_post_access(post_id, token_payload=token_payload):
+        PostRepo.delete_post(post_id)
+        return {"message": "post has been deleted"}
